5.4
Добавлен худ здоровья и система хила

+++ Assets/Scripts/Inventory/UsableItemInfo.cs
@@ -4,11 +4,11 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "InventoryUsableItemInfo", menuName = "Gameplay/Items/Create new UsableItemInfo")]
 public class UsableItemInfo : ScriptableObject, IUsableItemInfo
 {
-    [SerializeField] private float _damage;
+    [SerializeField] private int _damage;
     [SerializeField] private float _cooldown;
     [SerializeField] private float _range;
 
-    public float damage => _damage;
+    public int damage => _damage;
 
     public float coolDown => _cooldown;
